Modification fait dans Simulation : L'exercice 5 expliquait les bases d'un scheduler, il a donc été mit en tant qu'exercice 1, puis viennent tous les autres exercices appliquant les activator etc.. qu'on a dût renommé Déplacement de la description de ce qu'est un scheduler à l'exercice 1
