+++ src/conversation/ex02/Agent4.java
@@ -31,7 +31,7 @@ import madkit.kernel.Message;
 /**
  * Shows how agents exchange messages.
  * 
- * #jws conversation.ex02.sendReply2 jws#
+ * #jws conversation.ex03.sendReply2 jws#
  * 
  * To interact with other agents, agents need a form of communication. Doing so,the agents have MailBox, they can use it
  * to receive and send messages. Here, two agents are replying to the other one.
