+++ src/conversation/ex03/Agent5.java
@@ -28,7 +28,7 @@ import madkit.kernel.Message;
 /**
  * Agent5 works with Agent6, using a different role.
  * 
- * #jws conversation.ex03.ConversationID jws#
+ * #jws conversation.ex04.ConversationID jws#
  * 
  * To interact with other agents, agents need a form of communication. Doing so,the agents have MailBox, they can use it
  * to receive and send messages. Here, one agent is replying the other one, we will see the difference between
