+++ src/conversation/ex03/Agent6.java
@@ -32,7 +32,7 @@ import madkit.kernel.Message;
 /**
  * Shows how agents exchange messages.
  * 
- * #jws conversation.ex03.ConversationID jws#
+ * #jws conversation.ex04.ConversationID jws#
  * 
  * To interact with other agents, agents need a form of communication. Doing so,the agents have MailBox, they can use it
  * to receive and send messages. Here, one agent is replying to the other one, we will see the difference between
