+++ src/draft/ex05/LaunchingDifferentAgentTypes.java
@@ -6,7 +6,7 @@ import madkit.kernel.Agent;
  * Launching other kind of agents using executeThisAgent() is possible
  * using the madkit.kernel.Madkit.Option#launchAgents
  * 
- * 		#jws helloworld.ex05.LaunchingDifferentAgentTypes jws#
+ * 		#jws helloworld.ex02.LaunchingDifferentAgentTypes jws#
  */
 
 public class LaunchingDifferentAgentTypes extends Agent {
@@ -19,8 +19,8 @@ public class LaunchingDifferentAgentTypes extends Agent {
 
 
     public static void main(String[] args) {
-	executeThisAgent(2, true, "--launchAgents helloworld.ex03.ModifyingGUILocation,true,2");
+	executeThisAgent(2, true, "--launchAgents helloworld.ex04.ModifyingGUILocation,true,2");
 	//or
-	// executeThisAgent(2, true, Option.launchAgents.toString(),helloworld.ex03.ModifyingGUILocation.class.getName()+",2");
+	// executeThisAgent(2, true, Option.launchAgents.toString(),helloworld.ex04.ModifyingGUILocation.class.getName()+",2");
     }
 }
