+++ src/simulation/ex01/MyScheduler.java
@@ -1,55 +1,45 @@
 package simulation.ex01;
 
+import java.util.logging.Level;
+
 import madkit.gui.ConsoleAgent;
-import madkit.kernel.AbstractAgent;
 import madkit.kernel.Madkit;
 import madkit.kernel.Scheduler;
-import madkit.simulation.activator.GenericBehaviorActivator;
 
 /**
  * 
- *  #jws# simulation.ex01.MyScheduler #jws#
- *  #args# --launchAgents simulation.ex01.MyScheduler,true;madkit.gui.ConsoleAgent #args# 
- * 
- * In this first example, the goal is only to manage the execution of some agents. Two classes are needed : a
- * madkit.kernel.Scheduler that manages an madkit.kernel.Activator and a simulated agent class:
- * simulation.ex01.SimulatedAgent.
+ *  #jws# simulation.ex02.MyScheduler02 #jws#
+ * 	#args# --launchAgents simulation.ex02.MyScheduler02,true;madkit.gui.ConsoleAgent #args#
+ *
  *
  * The madkit.kernel.Scheduler class defines a generic threaded scheduler agent.
  * It holds a collection of activators. The default state of a scheduler is Scheduler.SimulationState.PAUSED
  *
- * The madkit.kernel.Activator class defines a tool for scheduling mechanism.
- * An activator is configured according to a community, a group and a role.
- * It could be used to activate a group of agents on a particular behavior
+ * Let us illustrate some basic features of a scheduler
  */
 
 public class MyScheduler extends Scheduler {
 
-    private GenericBehaviorActivator<AbstractAgent> activator1;
-
     @Override
     protected void activate() {
-
-	// 1 : create the simulation group
-	createGroup(SimulationModel.MY_COMMUNITY, SimulationModel.SIMU_GROUP);
-
-	// 2 : launch some simulated agents
-	for (int i = 0; i < 10; i++) {
-	    launchAgent(new SimulatedAgent());
-	}
-
-	// 3 : initialize the activator on the correct (1) CGR location and (2) behavior
-	activator1 = new GenericBehaviorActivator<AbstractAgent>(SimulationModel.MY_COMMUNITY, SimulationModel.SIMU_GROUP, SimulationModel.ROLE, "doIt");
-	addActivator(activator1);
-
-	// 4 : we are done, because Scheduler already defines a live method
-	// calling the execution of the activator. We will override it later.
-	// here we just slow down the simulation to not flood the console
+	// This makes a pause of 300 ms between two simulation steps and can be modified later in the code or
+	// using the scheduler GUI.
 	setDelay(300);
+
+	// Let us also display more information:
+	// the FINER log level of the Scheduler displays activation information
+	getLogger().setLevel(Level.FINER);
+
+	// The simulation will run until getGVT() >= 10 and then automatically quit
+	setSimulationDuration(10);
+
+	super.activate();
     }
 
     /**
      * A simple way of launching this scheduler
+     * 
+     * @param
      */
     public static void main(String[] args) {
 	new Madkit("--launchAgents", MyScheduler.class.getName() + ",true;" + ConsoleAgent.class.getName());
