+++ src/simulation/ex03/MyScheduler03.java
@@ -5,48 +5,38 @@ import madkit.kernel.AbstractAgent;
 import madkit.kernel.Madkit;
 import madkit.kernel.Scheduler;
 import madkit.simulation.activator.GenericBehaviorActivator;
-import simulation.ex01.SimulatedAgent;
-import simulation.ex01.SimulationModel;
 import simulation.ex02.SimulatedAgent02;
+import simulation.ex02.SimulationModel;
 
 /**
- * 
  *  #jws# simulation.ex03.MyScheduler03 #jws#
- * 	#args# --launchAgents simulation.ex03.MyScheduler03,true;madkit.gui.ConsoleAgent #args# 
+ *  #args# --launchAgents simulation.ex03.MyScheduler03,true;madkit.gui.ConsoleAgent #args#
  * 
- * Let us have more fun by adding another activator on another role to execute the other behavior of
- * {@link SimulatedAgent03}
+ * Let us have more fun by adding another simulated agent class with a different result for the doIt method and see what
+ * we now need. We need nothing else and just have to launch this new type of agent: Activators do not care about the
+ * exact type of the activated agents. Only the organizational position matters.
  */
 
 public class MyScheduler03 extends Scheduler {
 
-    protected GenericBehaviorActivator<AbstractAgent> activator1;
-    protected GenericBehaviorActivator<AbstractAgent> activator2;
+    private GenericBehaviorActivator<AbstractAgent> activator1;
 
     @Override
     protected void activate() {
-
 	// 1 : create the simulation group
 	createGroup(SimulationModel.MY_COMMUNITY, SimulationModel.SIMU_GROUP);
 
 	// 2 : launch some simulated agents
-	for (int i = 0; i < 10; i++) {
-	    launchAgent(new SimulatedAgent());
+	for (int i = 0; i < 20; i++) {
 	    launchAgent(new SimulatedAgent02());
 	    launchAgent(new SimulatedAgent03());
 	}
 
-	// 3 : initialize the activators
-	// by default, they are activated once each in the order they have been added
+	// 3 : initialize the activator
 	activator1 = new GenericBehaviorActivator<AbstractAgent>(SimulationModel.MY_COMMUNITY, SimulationModel.SIMU_GROUP, SimulationModel.ROLE, "doIt");
 	addActivator(activator1);
-	activator2 = new GenericBehaviorActivator<AbstractAgent>(SimulationModel.MY_COMMUNITY, SimulationModel.SIMU_GROUP, SimulationModel.ANOTHER_ROLE, "anotherBehavior");
-	addActivator(activator2);
 
 	setDelay(300);
-
-	// 4 : let us start the simulation automatically
-	setSimulationState(SimulationState.RUNNING);
     }
 
     /**
